+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   - fixed ParsleyFieldMultiple optional fields if not explicitely required
   - fixed ParsleyFieldMultiple trigger
   - fixed select UI auto-bind change on error (#537)
+  - fixed `asyncIsValid()` and `asyncValidate()` API to support `force` option
 
 ## 2.0.0-rc4